Everything is finally working with much more efficiency
