+++ frontend/web/src/routes/blog/[blogId]/blogtitle.svelte
@@ -11,8 +11,8 @@
 <div class="flex flex-row max-md:flex-col justify-between items-center gap-2">
 
 
-    {#if $blog.data.title}
-        <h3 class="scroll-m-20 text-3xl font-semibold tracking-tight">{$blog.data.title}</h3>
+    {#if $blog.title}
+        <h3 class="scroll-m-20 text-3xl font-semibold tracking-tight">{$blog.title}</h3>
     {:else if serverDown}
         <h3 class="scroll-m-20 text-3xl font-semibold tracking-tight text-red-400">Server Down. I am sorry...</h3>
     {:else if errormessage}
