+++ frontend/web/src/routes/blog/[blogId]/usertitle.svelte
@@ -6,10 +6,10 @@
     import { user } from "../../../stores/user.js";
   import { formatDate } from "$lib/util/dateFormat.js";
   import Skeleton from "$lib/components/ui/skeleton/skeleton.svelte";
-    $: url = `/user/${$user.data.username}`
+    $: url = `/user/${$user.username}`
 </script>
 
-{#if $user.data.username}
+{#if $user.username}
 <HoverCard.Root>
     <HoverCard.Trigger
     href={url}
@@ -17,21 +17,21 @@
     rel="noreferrer noopener"
     class="rounded-sm text-xl font-semibold tracking-tight underline-offset-4 hover:underline focus-visible:outline-2 focus-visible:outline-offset-8 focus-visible:outline-black"
     >
-    {$user.data.fullname}
+    {$user.fullname}
     </HoverCard.Trigger>
     <HoverCard.Content class="w-80">
     <div class="flex justify-between space-x-4">
         <Avatar.Root>
-        <Avatar.Image src="https://github.com/shadcn.png" />
-        <Avatar.Fallback>SK</Avatar.Fallback>
+        <Avatar.Image src={$user.avatar} />
+        <Avatar.Fallback>{$user.fullname.charAt(0)}</Avatar.Fallback>
         </Avatar.Root>
         <div class="space-y-1">
-            <h4 class="text-lg font-semibold">{$user.data.fullname}</h4>
-            <p class="text-sm">Software developer and a music nerd</p>
+            <h4 class="text-lg font-semibold">{$user.fullname}</h4>
+            <p class="text-sm">{$user.aboutme}</p>
         <div class="flex items-center pt-2">
             <Calendar class="mr-2 h-4 w-4 opacity-70" />
             <span class="text-xs text-muted-foreground">
-            Joined {formatDate($user.data.createdAt)}
+            Joined {formatDate($user.createdAt)}
             </span>
         </div>
         </div>
