--- frontend/web/src/routes/user/+page.svelte
@@ -1,7 +1,0 @@
-<script>
-    import { onMount } from "svelte";
-
-    onMount( async() => {
-        window.location.href = '/user/lua';
-    })
-</script>
