+++ frontend/web/src/routes/user/[username]/avatarDialogue.svelte
@@ -7,42 +7,48 @@
         import * as Dialog from "$lib/components/ui/dialog/index.js";
     import { user } from "../../../stores/user";
     import { basic } from "../../../stores/basic";
-  import Skeleton from "$lib/components/ui/skeleton/skeleton.svelte";
+    import Skeleton from "$lib/components/ui/skeleton/skeleton.svelte";
+    import Reload from "svelte-radix/Reload.svelte";
 
     let editImageContent = false
     let dialogOpen = false
     $: if (dialogOpen) {
         editImageContent = false;
     }
+    
 
     let selectedFiles: any = [];
 
     // YET TO IMPLEMENT LOADING ANIMATION
 
-    // let loading = false
-    // let disabled = ""
-    // $: if (loading) {
-    //     disabled = "disabled"
-    // } else {
-    //     disabled = ""
-    // }
+    let loading = false
+    let disabled = ""
+    $: if (loading) {
+        disabled = "disabled"
+    } else {
+        disabled = ""
+    }
     async function removeImage() {
+        loading = true
         try {
             const res = await fetch("/api/v1/users/remove-avatar", {
                 method: "POST"
             })
             const data = await res.json();
             console.log("Removal successful:", data);
-            user.set(data)
+            user.set(data.data)
             editImageContent = false
         } catch (error) {
             console.error("Error removing image:", error);
             alert("Failed to upload avatar.");
             editImageContent = false
+        } finally {
+            loading= false
         }
     }
 
     async function saveImage() {
+        loading = true
         if (selectedFiles.length === 0) {
             alert("No file selected.");
             return;
@@ -63,12 +69,14 @@
             const data = await res.json();
             console.log("Upload successful:", data);
 
-            user.set(data)
+            user.set(data.data)
             editImageContent = false
         } catch (error) {
             console.error("Error uploading image:", error);
             alert("Failed to upload avatar.");
             editImageContent = false
+        } finally {
+            loading= false
         }
     }
 </script>
@@ -76,8 +84,8 @@
 <Dialog.Root bind:open={dialogOpen}>
     <Dialog.Trigger
     >
-        {#if $user.data?.avatar !== ""}
-            <img id="avatar" src={$user.data?.avatar} alt="avatar">
+        {#if $user.avatar !== ""}
+            <img id="avatar" src={$user.avatar} alt="avatar">
         {:else}
             <Skeleton class="w-[150px] h-[150px] rounded-xl"></Skeleton>
         {/if}
@@ -92,8 +100,8 @@
                 </Dialog.Description>
             </Dialog.Header>
             <div class="grid gap-4 py-4">
-                {#if $user.data.avatar !== ""}
-                <img id="avatarDisplay" src={$user.data.avatar} alt="avatarDisplay">
+                {#if $user.avatar !== ""}
+                <img id="avatarDisplay" src={$user.avatar} alt="avatarDisplay">
                 {:else}
                 <Skeleton class="w-full aspect-square"></Skeleton>
                 {/if}
@@ -101,8 +109,13 @@
 
             {#if $basic.sameUser === 1}
             <Dialog.Footer>
-                <Button type="submit" variant="outline">Remove Image</Button>
-                <Button type="submit" on:click={() => {editImageContent = true}}>Edit Image</Button>
+                <Button type="submit" variant="outline" on:click={removeImage} {disabled}>
+                    {#if loading}
+                        <Reload class="mr-2 h-4 w-4 animate-spin" ></Reload>
+                    {/if}
+                    Remove Image
+                </Button>
+                <Button type="submit" on:click={() => {editImageContent = true}} {disabled}>Edit Image</Button>
             </Dialog.Footer>
             {/if}
         {:else}
@@ -114,8 +127,13 @@
             </Dialog.Header>
             <FileInput on:filechange={(event) => (selectedFiles = event.detail.files)} />
             <Dialog.Footer>
-                <Button type="submit" on:click={() => {editImageContent = false}} variant="outline">Cancel</Button>
-                <Button type="submit" on:click={saveImage}>Save Image</Button>
+                <Button type="submit" on:click={() => {editImageContent = false}} variant="outline" {disabled}>Cancel</Button>
+                <Button type="submit" on:click={saveImage} {disabled}>
+                    {#if loading}
+                        <Reload class="mr-2 h-4 w-4 animate-spin" ></Reload>
+                    {/if}
+                    Save Image
+                </Button>
             </Dialog.Footer>
         {/if}
     </Dialog.Content>
