+++ frontend/web/src/routes/user/[username]/search.svelte
@@ -2,13 +2,13 @@
     import Input from "$lib/components/ui/input/input.svelte";
     import {Separator} from "$lib/components/ui/separator/index"
     import { MagnifyingGlass } from "svelte-radix"
-    import { user } from "../../../stores/user"
     import { onDestroy, onMount } from "svelte"
     import Button from "$lib/components/ui/button/button.svelte";
-    import {Person} from "svelte-radix"
     import Settings from "./settings.svelte";
+  import { basic } from "../../../stores/basic";
 
-    export let loggedinFlag = false
+    let loggedinFlag = $basic.sameUser === 3
+    $: loggedinFlag = $basic.sameUser !== 3
 
     let searchQuery = ""
     let searchResults = []
@@ -60,7 +60,7 @@
                 <div class="h-10 flex justify-center items-center bg-black rounded-es-xl rounded-ss-xl aspect-square border-[1px] border-r-0">
                     <MagnifyingGlass size="30"></MagnifyingGlass>
                 </div>
-                <Input on:input={search} bind:value={searchQuery} class="w-full rounded-none rounded-ee-xl rounded-se-xl" placeholder="Search"/>
+                <Input on:input={search} bind:value={searchQuery} class="w-full rounded-none rounded-ee-xl rounded-se-xl" placeholder="Search Users"/>
             </div>
 
 
@@ -95,11 +95,7 @@
 
 
 
-        {#if loggedinFlag}
-<!--            <div class=" aspect-square h-10 bg-zinc-950 border-[1px] rounded-full flex justify-center items-center">-->
-<!--                <Person></Person>-->
-<!--            </div>-->
-        {:else}
+        {#if !loggedinFlag}
             <Button variant="secondary" on:click={() => {window.open("/login", "_self")}}>Login</Button>
         {/if}
 
@@ -113,15 +109,11 @@
             <Input class="w-full rounded-none rounded-ee-xl rounded-se-xl max-[580px]:hidden" placeholder="Search (Coming Soon)"/>
         </div>
 
-        {#if loggedinFlag}
-<!--            <div class=" aspect-square h-10 bg-zinc-950 border-[1px] rounded-full flex justify-center items-center">-->
-<!--                <Person></Person>-->
-<!--            </div>-->
-        {:else}
+        {#if !loggedinFlag}
             <Button variant="secondary" on:click={() => {window.open("/login", "_self")}}>Login</Button>
         {/if}
     </div>
 
     <div class="min-w-1"></div>
-    <Settings sameUser={loggedinFlag}></Settings>
+    <Settings></Settings>
 </div>
