+++ frontend/web/src/routes/user/[username]/topbar.svelte
@@ -8,31 +8,18 @@
     import {Separator} from "$lib/components/ui/separator/index"
     import { Skeleton } from "$lib/components/ui/skeleton";
     import { basic } from "../../../stores/basic";
+  import { waitUntilNotEqual } from "$lib/util/waitTillNotEqual";
 
-    let sameUser = 2
-    // 0 => not same user
-    // 1 => same user
-    // 2 => loading
-    // 3 => not logged in (not same user)
-    let loggedin_confirm = false
     let subscribed_or_not = false
     let subscriptionId
+    let sameUserValue = 2
+    
 
     onMount(async () => {
-        sameUser = 2
-        loggedin_confirm = false
 
-        try {
-            const url = new URL(window.location.href);
-            const username = url.pathname.split('/').pop(); // Get the last part of the URL
-
-            const res = await fetch(`/api/v1/users/get-user/${username}`);
-            const data = await res.json()
-            user.set(data);
-            console.log($user);
-        } catch (error) {
-
-        }
+        console.log("Waiting till user is not null");
+        await waitUntilNotEqual(() => $user._id, null)
+        console.log("User is not null");
 
         try {
             const res = await fetch(`/api/v1/users/loggedin-confirm`, {
@@ -41,30 +28,26 @@
                 });
             const data = await res.json();
             console.log(data)
-            if (res.status == 401) {
-                basic.set({sameUser: 3})
-                sameUser = 3
+            if (res.status === 401) {
+                sameUserValue = 3
+                basic.set({sameUser: sameUserValue})
                 return
             }
-            if (data.data?._id === $user.data?._id) {
-                basic.set({sameUser: 1})
-                sameUser = 1
+            if (data.data?._id === $user._id) {
+                sameUserValue= 1
             } else {
-                console.log(data.data?._id, "\n" , $user.data?._id)
-                basic.set({sameUser: 0})
-                sameUser = 0
+                // console.log(data.data?._id, "\n" , $user._id)
+                sameUserValue = 0
             }
             
         } catch (error) {
             console.log(error);
             
-            sameUser = 2
-            basic.set({sameUser: 2})
-            loggedin_confirm = false
+            sameUserValue = 2
         }
 
         try {
-            if (sameUser == 0) {
+            if (sameUserValue == 0) {
                 const url = new URL(window.location.href);
                 let username = url.pathname.split('/').pop(); // Get the last part of the URL
                 
@@ -93,18 +76,21 @@
         } catch (error) {
             subscribed_or_not = false
             console.log(error);
+        } finally {
+            basic.set({sameUser: sameUserValue})
         }
+        
     });
 
     const subscribe = async () => {
         try {
-            console.log($user.data.username);
+            console.log($user.username);
             
             const res = await fetch(`/api/v1/subscription/subscribe`, {
                 method: "POST",
                 headers: { 'Content-Type': 'application/json' },
                 body: JSON.stringify({
-                    blogger_username: $user.data.username
+                    blogger_username: $user.username
                 })
             })
             const data = await res.json()
@@ -117,6 +103,7 @@
                 subscribed_or_not = false
             }
         } catch (error) {
+            console.log(error);
             subscribed_or_not = false
         }
     }
@@ -153,22 +140,13 @@
 
         <div id="right_container" class="flex flex-row items-center grow justify-between max-sm:flex-col max-sm:items-start max-sm:gap-2">
             <div id="names_container">
-                {#if $user && $user.data}
-                    <h2 class="scroll-m-20 text-3xl font-semibold tracking-tight">{$user.data.fullname}</h2>
-                    <p>
-                        @{$user.data.username}
-                    </p>
-                {/if}
+                <h2 class="scroll-m-20 text-3xl font-semibold tracking-tight">{$user.fullname}</h2>
+                <p>
+                    @{$user.username}
+                </p>
                 <div class="h-4"></div>
 
 
-
-                <!-- {#if $subscribers && $subscribers.data && $subscribers.data && $subscribers.data.userSubscribersCount}
-                    <p class="underline underline-offset-2"><span class="cursor-pointer"><span class=" text-2xl">{$subscribers.data.userSubscribersCount}</span> Subscribers</span></p>
-                {/if}
-                {#if $subscriptions && $subscriptions.data && $subscriptions.data && $subscriptions.data.userSubscriptionsCount}
-                    <p class="underline underline-offset-2"><span class="cursor-pointer"><span class=" text-2xl">{$subscribers.data.userSubscriptionsCount}</span> Subscriptions</span></p>
-                {/if} -->
 
                 <p>
                     {#if $subscribers && $subscribers.data && $subscribers.data && $subscribers.data.userSubscribersCount}
@@ -186,15 +164,15 @@
             </div>
             <!-- <div class="grow"></div> -->
             <div class="flex flex-col justify-center items-center">
-                {#if sameUser === 0}
+                {#if $basic.sameUser === 0}
                     {#if subscribed_or_not}
                     <Button variant="secondary" class="w-24" on:click={unsubscribe}>Unsubscribe</Button>
                     {:else}
                     <Button variant="default" class="w-24" on:click={subscribe}>Subscribe</Button>
                     {/if}
-                {:else if sameUser === 1}
+                {:else if $basic.sameUser === 1}
                     <Button variant="secondary" class="w-24" on:click={() => {window.open('/blog/create-blog', "_self")}}>Create Blog</Button>
-                {:else if sameUser === 2}
+                {:else if $basic.sameUser === 2}
                     <Skeleton class="w-24 h-10 rounded-md"></Skeleton>
                 {:else}
                     <Button variant="default" class="w-24" on:click={() => {window.open("/login", "_self")}}>Subscribe</Button>
