--- frontend/web/src/routes/userNew/[username]/banner.svelte
@@ -1,9 +1,0 @@
-<div class="bgimg_banner w-full h-[150px] rounded-[20px] border-2 bg-black"></div>
-
-<style>
-    .bgimg_banner {
-        background-image: url("https://wallpapercave.com/wp/xIzYo2d.jpg");
-        background-size: cover;
-        background-position: 50%;
-    }
-</style>
