--- frontend/web/src/routes/userNew/[username]/rightBoxContent.svelte
@@ -1,74 +1,0 @@
-<script>
-  import Listcontent from "$lib/components/my_components/listcontent.svelte";
-import Socialslist from "$lib/components/my_components/socialslist.svelte";
-  import Button from "$lib/components/ui/button/button.svelte";
-  import Skeleton from '$lib/components/ui/skeleton/skeleton.svelte';
-  import { settingSheet } from "../../../stores/sheets";
-  import Interests from "./interests.svelte";
-
-    // export let aboutMe = `Lorem ipsum dolor sit amet consectetur adipisicing elit. Iure, esse cupiditate? Nobis quae voluptatum ab facere ratione. Est, reprehenderit similique aliquid, recusandae et veritatis harum quod quidem eaque maiores error.
-    // Autem vero pariatur quidem ipsa expedita laboriosam illum debitis nobis animi! Modi ea, itaque corrupti vero corporis nisi. Amet soluta possimus facere minus eum error sed quasi nisi beatae aspernatur!`
-    export let instagram
-    export let linkedin = ""
-    export let facebook
-    export let github
-    export let aboutMe
-    export let loggedinFlag = false
-    export let noAboutMe = false
-    export let noSocials = false
-    export let noInterests = false
-</script>
-
-
-{#if loggedinFlag}
-<Button variant="secondary" class="mb-2 w-fit m-2 absolute top-0 right-0"   on:click={() => { settingSheet.set({...$settingSheet, openEditInfo: true}); }} >Edit Info</Button>
-{/if}
-
-<h1 class="scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-4xl">
-    About me:
-</h1>
-{#if aboutMe}
-    <p class="mt-1">{aboutMe}</p>
-{:else if noAboutMe}
-    <p class="mt-1 text-muted-foreground">User has not added about me</p>
-{:else}
-    <div class="flex flex-col gap-1">
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-        <Skeleton class="w-full h-5 rounded-md"></Skeleton>
-    </div>
-{/if}
-<hr class="my-4">
-<h1 class="scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-4xl">
-    Socials:
-</h1>
-<div class="relative h-[80px]">
-    <div class="flex w-full absolute left-0 -translate-x-0.5 items-center gap-[8px] py-2 overflow-x-scroll min-h-[80px]">
-        {#if noSocials}
-            <Listcontent title={"No socials added"} date={": ("}></Listcontent>
-        {:else}
-            {#if instagram}
-                <Socialslist title={"Instagram"} socialId={instagram} imagePath={"/assets/instagram_logo.png"}></Socialslist>
-            {/if}
-            {#if facebook}
-                <Socialslist title={"Facebook"} socialId={facebook} imagePath={"/assets/facebook_logo.png"}></Socialslist>
-            {/if}
-            {#if github}
-                <Socialslist title={"Github"} socialId={github} imagePath={"/assets/github_logo.png"}></Socialslist>
-            {/if}
-            {#if linkedin}
-                <Socialslist title={"LinkedIn"} socialId={linkedin} imagePath={"/assets/linkedin_logo.png"}></Socialslist>
-            {/if}
-            {#if !linkedin && !facebook && !github && !instagram}
-                <Skeleton class=" min-w-60 h-[3.75rem] rounded-[15px]"></Skeleton>
-                <Skeleton class=" min-w-60 h-[3.75rem] rounded-[15px]"></Skeleton>
-                <Skeleton class=" min-w-60 h-[3.75rem] rounded-[15px]"></Skeleton>
-            {/if}
-        {/if}
-    </div>
-</div>
-<br>
-<Interests noInterests={noInterests}></Interests>
